Rearrange packages
Moves `./cmd/frontend/internal/search/zoekt/zoekt.go` to
`./internal/search/zoekt` so that we're not importing between services.

Also moves `HandleFilePathPatterns` into searcher. For now, the frontend
imports from searcher, but structural search is removed from the
frontend, that dependency will be removed.
